requirements.txt
Flask
pytesseract
Pillow
PyMuPDF
pdf2image
